Extract Haru FFI code from php docs Pull Request
On an attempt to revive the PHP PDF documentation, I filed
https://github.com/php/phd/pull/69.

The Pull Request above proposes substituting the now unmaintained
libharu PECL extension with a minimal libharu FFI package which is
compatible with the PECL extension.

We now start extracting the FFI package in case this is useful for other
users.